+++ .github/workflows/_e2e_test.yaml
@@ -902,7 +902,7 @@ jobs:
   e2e_310p:
     name: 310p singlecard
     runs-on: linux-aarch64-310p-1
-    if: ${{ (inputs.contains_310 || inputs.p310_tests != '') && false }}
+    if: ${{ inputs.contains_310 || inputs.p310_tests != '' }}
     container:
       image: ${{ inputs.image_310p }}
       env:
@@ -1014,7 +1014,7 @@ jobs:
   e2e_310p-4cards:
     name: 310p multicards 4cards
     runs-on: linux-aarch64-310p-4
-    if: ${{ inputs.contains_310 && false }}
+    if: ${{ inputs.contains_310 }}
     container:
       image: ${{ inputs.image_310p }}
       env:
